+++ mav_msgs/include/mav_msgs/eigen_mav_msgs.h
@@ -5,6 +5,8 @@
  * Copyright 2015 Markus Achtelik, ASL, ETH Zurich, Switzerland
  * Copyright 2015 Helen Oleynikova, ASL, ETH Zurich, Switzerland
  * Copyright 2015 Mina Kamel, ASL, ETH Zurich, Switzerland
+ * Copyright 2020 Ria Sonecha, MIT, USA
+ * Copyright 2020 Giuseppe Silano, University of Sannio in Benevento, Italy
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -25,6 +27,7 @@
 #include <Eigen/Eigen>
 #include <deque>
 #include <iostream>
+#include <time.h>
 
 #include "mav_msgs/common.h"
 
@@ -319,6 +322,41 @@ struct EigenOdometry {
   }
 };
 
+struct EigenDroneState {
+  EigenDroneState()
+      : timestamp_ns(ros::Time::now()),
+        position_W(Eigen::Vector3f::Zero()),
+        orientation_W_B(Eigen::Quaterniond::Identity()),
+        velocity(Eigen::Vector3f::Zero()),
+        acceleration(Eigen::Vector3f::Zero()),
+        angular_velocity_B(Eigen::Vector3f::Zero()),
+        angular_acceleration_B(Eigen::Vector3f::Zero()) {}
+
+  EigenDroneState(const Eigen::Vector3f& _position,
+                  const Eigen::Quaterniond& _orientation,
+                  const Eigen::Vector3f& _velocity,
+                  const Eigen::Vector3f& _acceleration,
+                  const Eigen::Vector3f& _angular_velocity_body,
+                  const Eigen::Vector3f& _angular_acceleration_body)
+      : position_W(_position),
+        orientation_W_B(_orientation),
+        velocity(_velocity),
+        acceleration(_acceleration),
+        angular_velocity_B(_angular_velocity_body),
+        angular_acceleration_B(_angular_acceleration_body) {}
+
+  EIGEN_MAKE_ALIGNED_OPERATOR_NEW
+  ros::Time
+      timestamp_ns;  // Time since epoch, negative value = invalid timestamp.
+  Eigen::Vector3f position_W;
+  Eigen::Quaterniond orientation_W_B;
+  Eigen::Vector3f velocity;
+  Eigen::Vector3f acceleration;
+  Eigen::Vector3f angular_velocity_B;
+  Eigen::Vector3f angular_acceleration_B;
+
+};
+
 // TODO(helenol): replaced with aligned allocator headers from Simon.
 #define MAV_MSGS_CONCATENATE(x, y) x##y
 #define MAV_MSGS_CONCATENATE2(x, y) MAV_MSGS_CONCATENATE(x, y)
@@ -335,6 +373,7 @@ MAV_MSGS_MAKE_ALIGNED_CONTAINERS(EigenTrajectoryPoint)
 MAV_MSGS_MAKE_ALIGNED_CONTAINERS(EigenRollPitchYawrateThrust)
 MAV_MSGS_MAKE_ALIGNED_CONTAINERS(EigenRollPitchYawrateThrustCrazyflie)
 MAV_MSGS_MAKE_ALIGNED_CONTAINERS(EigenOdometry)
+MAV_MSGS_MAKE_ALIGNED_CONTAINERS(EigenDroneState)
 }
 
 #endif  // MAV_MSGS_EIGEN_MAV_MSGS_H